create module eqparams and reorganize class's structure
